pyproject.toml
[project]
name = "llm-quant-lab"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
  "numpy>=1.26",
  "pandas>=2.0",
  "scipy>=1.11",
  "typer[all]>=0.12",
  "rich>=13",
  "pyyaml>=6",
  "matplotlib>=3.8",
  "plotly>=5.20",
  "tqdm>=4.66",
  "python-dotenv>=1.0",
  "transformers>=4.43,<4.44",
  "torch>=2.2",
  "accelerate>=0.30",
  "bitsandbytes>=0.43",
  "psutil>=5.9",
  "datasets>=2.19",
]

[project.scripts]
lab = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
  "ipykernel>=6",
  "jupyter>=1",
  "ruff>=0.4",
  "pytest>=8",
  "pytest-cov>=5",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
